Generic servicelerle ilgili bir kaç düzeltme vebakım planının eklenmesi

+++ ApplicationCore/Concrete/DeviceService.cs
@@ -7,6 +7,7 @@ using Domain.Entites;
 using Domain.Identity;
 using Domain.Repository;
 using Microsoft.AspNetCore.Http;
+using Microsoft.EntityFrameworkCore.Metadata.Internal;
 using Microsoft.Identity.Client;
 using System;
 using System.Collections.Generic;
@@ -28,60 +29,59 @@ namespace ApplicationCore.Concrete
             _mapper = mapper;
             _readRepository= readRepository;
         }
-        public async Task<CreateDeviceDto> CreateDevice(CreateDeviceDto device, Expression<Func<Devices, bool>> method, params Expression<Func<Devices, object>>[] includes)
-        {
-           var entity=_mapper.Map<Devices>(device);
-           var dto = await _baseService.AddAsync(entity,null,method,includes);
-           return _mapper.Map<CreateDeviceDto>(dto);
+        public async Task<CreateDeviceDto> CreateDevice(CreateDeviceDto device, params Expression<Func<Devices, object>>[] includes)
+        {   Expression<Func<Devices,bool>> method=x=>x.serialNo==device.serialNo;
+            var entity=await _readRepository.GetSingleAsync(method);
+            if (entity == null)
+            {
+                var models = _mapper.Map<Devices>(device);
+                var dto = await _baseService.AddAsync(models, null, method, includes);
+                return _mapper.Map<CreateDeviceDto>(dto);
+            }
+            return null;
 
         }
-        public async Task<UpdateDeviceDto> UpdateDevice(UpdateDeviceDto models,Expression<Func<Devices,bool>> method,params Expression<Func<Devices, object>>[] includes)
+        public async Task<UpdateDeviceDto> UpdateDevice(UpdateDeviceDto models,Guid id,params Expression<Func<Devices, object>>[] includes)
         {
+            Expression<Func<Devices,bool>> method=x=>x.Id==id;
             var entity = await _readRepository.GetSingleAsync(method);
-            if (entity == null) 
-            {
-                throw new ArgumentNullException("Böyle bir entity yok");
-            }
+            if (entity == null) return null;
             var model = _mapper.Map(models, entity);
             var result = await _baseService.UpdateAsync(model, method,includes);
-            if (result==null)
-            {
-                throw new ArgumentNullException("Dto hatalı döndü");
-            }
             return _mapper.Map<UpdateDeviceDto>(result);
         }
         public async Task<DeleteDeviceDtos> DeleteDevice(Guid id)
-        {
-            var result=await _baseService.DeleteAsync(id);
-
+        {   var entity=await _readRepository.GetByIdAsync(id);
+            if (entity != null)
+            {
+                var result = await _baseService.DeleteAsync(id);
                 var model = _mapper.Map<DeleteDeviceDtos>(result);
-                return model;
+            }
+            return _mapper.Map<DeleteDeviceDtos>(entity);
  
         }
         public async Task<List<GetDeviceDto>> GetAllDevice()
         {
             var result = await _baseService.GetAllAsync();
             var users=new List<GetDeviceDto>();
-            if (result == null)
+
+            foreach (var models in result)
             {
-                throw new ArgumentNullException("Cihaz yok");
-            }
-            foreach(var models in result)
-            {
-                var dto=_mapper.Map<GetDeviceDto>(models);
-                users.Add(dto);
+             var dto = _mapper.Map<GetDeviceDto>(models);
+             users.Add(dto);
             }
             return users;
         }
-        public async Task<GetDeviceDto> GetByIdDevice(Expression<Func<Devices, bool>> filtre, params Expression<Func<Devices, object>>[] includes)
+        public async Task<GetDeviceDto> GetByIdDevice(Guid id, params Expression<Func<Devices, object>>[] includes)
         {
-            var result = await _baseService.GetByIdAsync(filtre, includes);
-            if (result == null) 
+            Expression<Func<Devices, bool>> filtre = x => x.Id == id;
+            var entity = await _readRepository.GetSingleAsync(filtre);
+            if (entity != null)
             {
-                throw new ArgumentNullException("Böyle bir cihaz sistemimizde bulunmamaktadır");
-
+                var result = await _baseService.GetByIdAsync(filtre, includes);
+                return _mapper.Map<GetDeviceDto>(result);
             }
-            return _mapper.Map<GetDeviceDto>(result);
+            return _mapper.Map<GetDeviceDto>(entity);
           
         }
